refactor: split filter-helpers.ts into focused modules (#180)
Extract filter-repost-status, filter-display, filter-group-ops,
filter-legacy, and filter-types from 601-line filter-helpers.
Re-exports preserve existing import paths. No behavior changes.
